toolkit/devtools/server/tests/mochitest/test_css-logic.html

<!DOCTYPE HTML>
<html>



<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
Cu.import("resource://gre/modules/devtools/Loader.jsm");
const {Promise: promise} = Cu.import("resource://gre/modules/Promise.jsm", {});

const {CssLogic} = devtools.require("devtools/styleinspector/css-logic");
Cu.importGlobalProperties(['CSS']);

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

addTest(function findAllCssSelectors() {
  var nodes = document.querySelectorAll('*');
  for (var i = 0; i < nodes.length; i++) {
    var selector = CssLogic.findCssSelector(nodes[i]);
    var matches = document.querySelectorAll(selector);

    is(matches.length, 1, 'There is a single match: ' + selector);
    is(matches[0], nodes[i], 'The selector matches the correct node: ' + selector);
  }

  runNextTest();
});

addTest(function findCssSelectorNotContainedInDocument() {

  var unattached = document.createElement("div");
  unattached.id = "unattached";
  try {
    CssLogic.findCssSelector(unattached);
    ok (false, "Unattached node did not throw")
  } catch(e) {
    ok(e, "Unattached node throws an exception");
  }

  var unattachedChild = document.createElement("div");
  unattached.appendChild(unattachedChild);
  try {
    CssLogic.findCssSelector(unattachedChild);
    ok (false, "Unattached child node did not throw")
  } catch(e) {
    ok(e, "Unattached child node throws an exception");
  }

  var unattachedBody = document.createElement("body");
  try {
    CssLogic.findCssSelector(unattachedBody);
    ok (false, "Unattached body node did not throw")
  } catch(e) {
    ok(e, "Unattached body node throws an exception");
  }

  runNextTest();
});

addTest(function findCssSelector() {

  let data = [
    "#one",
    "#" + CSS.escape("2"),
    ".three",
    "." + CSS.escape("4"),
    "#find-css-selector > div:nth-child(5)",
    "#find-css-selector > p:nth-child(6)",
    ".seven",
    ".eight",
    ".nine",
    ".ten",
    "div.sameclass:nth-child(11)",
    "div.sameclass:nth-child(12)",
    "div.sameclass:nth-child(13)",
    "#" + CSS.escape("!, \", #, $, %, &, ', (, ), *, +, ,, -, ., /, :, ;, <, =, >, ?, @, [, \\, ], ^, `, {, |, }, ~"),
  ];

  let container = document.querySelector("#find-css-selector");
  is (container.children.length, data.length, "Container has correct number of children.");

  for (let i = 0; i < data.length; i++) {
    let node = container.children[i];
    is (CssLogic.findCssSelector(node), data[i], "matched id for index " + (i-1));
  }

  runNextTest();
});

addTest(function getBackgroundImageUriFromProperty() {

  let data = [
    ["background: url(foo.png);", "foo.png"],
    ["background: url(\"foo.png\") ", "foo.png"],
    ["background: url('foo.png') ; ", "foo.png"],
    ["background: foo.png", null],
    ["background: url()", ""],
  ];

  for (let i = 0; i < data.length; i++) {
    let prop = data[i][0];
    let result = data[i][1];
    is (CssLogic.getBackgroundImageUriFromProperty(prop), result,
      "Background image matches for index " + i);
  }

  runNextTest();
});

  </script>
</head>
<body>
  <div id="find-css-selector">
    <div id="one"></div> 
    <div id="2"></div> 
    <div class="three"></div> 
    <div class="4"></div> 
    <div attr="5"></div>  
    <p></p> 
    <div class="seven seven"></div> 
    <div class="eight eight2"></div> 

    
    <div class="nine" id="nine-and-ten"></div>
    <div class="ten" id="nine-and-ten"></div>

    
    <div class="sameclass" id="11-12-13"></div>
    <div class="sameclass" id="11-12-13"></div>
    <div class="sameclass" id="11-12-13"></div>

    
    <div id="!, &quot;, #, $, %, &amp;, ', (, ), *, +, ,, -, ., /, :, ;, <, =, >, ?, @, [, \, ], ^, `, {, |, }, ~"></div>
  </div>
</body>
</html>
